Allow more photos, changed field name
We now allow a user or API call to add up to three photos to a report.  Some work still must be done to intelligently handle using up the empty photo slots first during an update, but the majority of the work of getting three photos added to a report is there.

Change the "type" field to "lot_type"  "type" is a reserved Python word, and having the field name as the same was shadowing the type command, making my debugging a real challenge.  "Why does it think type is a  unicode string instead of a built-in function?"